Add files via upload
Section 1
Summary
Today we will write some programs involving text file input and output. We will be importing some new libraries that contain methods necessary for making file I/O work.

Part 1
In a class called Stats (template is provided), given a text file of doubles (see this lab’s Canvas page for a sample file), write a program that determines the average, the maximum, the minimum and how many numbers are in the following buckets: less than 0, between 0 (inclusive) and 100 (exclusive), and greater than or equal to 100.

Your program should print the following message to another file called “fileOut.txt”:


    Statistics for the numbers in fileIn.txt:
    average: <average>
    max: <max>
    min: <min>

    There are <negNum> negative numbers, <btw0and100> numbers between 0 (inclusive) and 100 (exclusive), and <geq100> numbers that are greater than or equal to 100.
  
(Optional Challenge - work on it after completing the rest of the lab!)

Write a program that generates a text file that can be used as input for the problem described above. Use your knowledge of file output, loops and the random method from the Math library. How random and in what range the numbers should be it totally up to you!

Part 2
In a class called Diary (template is provided), write a program that prompts users to enter the date as three integers separated by spaces (i.e mm dd yyyy) then it prompts them enter as many lines of prose they wish (for their to-do’s list or diary entry). Your program should store their entries at the end of a file called “diaryLog.txt” and old data should NOT be erased. Each entry should be formatted as follows (example for a diary entry entered on 7/14/2017):


    ...old stuff.
    Date: 07/14/2017
    Dear diary, today we had a midterm. I think it went ok, ..
  
Part 3
Write a program Advice.java that gives and takes advice on program writing. The program starts by writing a piece of advice to the screen and asking the user to type in a different piece of advice. The program then ends.

The next person to run the program receives the advice given by the person who last ran the program.

The advice is kept in a text file advice.txt and the content of the file changes after each run of the program. Furthermore, you should have another text file called adviceLog.txt, which starts off blank but with each run of the program it logs the advice that was given (without deleting the previous advices). You should allow theuser to type in advice of any length so that it can be any number of lines long. The user is told to end his orher advice by pressing the Return key two times. Your program can then test to see that it has reached the end of the input by checking to see when it reads an empty string "".

Input file: advice.txt

Section 2
Template: Lab6.javaPreview the document

Summary
Today we will write some programs involving String objects. Strings are widely used not only in Java, but in many other programming languages, as well. You can think of a string as a sequence of characters; strings and are treated as objects in Java and Java platform provides the String class to create and manipulate strings.

Part 1
Write a method that capitalizes the first letter of each word in a sentence, named capitalizeFirstLetters. This method prints a capitalized version of a String to the screen -that is, it changes the first letter of each word in the string to upper case (if it is not already upper case). So "I really enjoy attending lab!" becomes "I Really Enjoy Attending Lab!". The string to be printed should be a parameter to the method, and the method should have a void return. You should test your method in main by getting a line of input from the user and applying the method to it.


        public static void capitalizeFirstLetters(String input) {
          //TODO: parse input for words and capitalize each one.
        }
      
Hint: You may want to browse the Java String docs for useful methods on strings, such as split().

Hint: You can usea new Scanner to get next() words out of a String.

Hint: StringTokenizer is a class that can do the same thing.

Write a method called nameInitials that takes one String as argument, pertaining to somebody's full name and returns a String of the name's initials. Usage example,


        String initials = nameInitials("Bertrand Arthur William Russell");
        System.out.println(initials);         //should print B.A.W.R.
      
Optional

Write a method called letterCount that takes two String arguments, one containing some text and the other containing a single letter. This method should then compute and return the count of all of the occurrences of that letter in the given text. Usage example,


        int countMs;
        countMs = letterCount("I am the very model of a modern major general", "m");
        System.out.println(countMs);       //should print 4
      
 

Part 2
Write a method called lexLargest that takes a string object argument containing some text. Your method should return the lexicographical largest word in the String object (that is, the one that would appear latest in the dictionary).


        String large = lexLargest ("I am the very model of a modern major general");
        System.out.println(large);   // should print : very
      
Write a method called largestBy that takes a string object argument containing some text. Your method should return the largest-by-length word in the String object. In case of ties, return the first occurrence of the largest word. Note the return type is String. Example,


        String large ;
        large = largestBy("I am the very model of a modern major general");
        System.out.println(large);                   // should print general
      
Optional

Now, using above two methods, see if you can combine them in one single method and still produce the same outputs? NOTE: use an integer to determine which algorithm your code will execute, e.g., String largestByAction(String inputText, int action) If integer action is 1, then run first algorithm and if integer action is 2, then run the 2nd algorithm. Also, if the integer action is anything other than 1 or 2, return the empty string "".

Part 3
Write a Java program named CharsIndex to get the index of all the characters of the alphabet. Ask user to input a string.The sample output was produced by this string: "The quick brown fox jumps over the lazy dog."

Sample Output:


    a  b  c  d e  f  g h i  j
    =========================
    36 10 7 40 2 16 42 1 6 20

    k  l  m  n  o  p q  r  s  t
    ===========================
    8 35 22 14 12 23 4 11 24 31

    u  v  w  x  y  z
    ================
    5 27 13 18 38 37